pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shuck-convert"
version = "1.0.0"
description = "Convert between document formats — DOCX/PDF to Markdown, Markdown to DOCX."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shan" },
]
keywords = ["document", "conversion", "markdown", "pdf", "docx", "mcp", "ai-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "fastmcp>=2.0",
    "python-docx>=0.8.11",
    "pdfplumber>=0.9.0",
    "PyMuPDF>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/Shan-Zhu/shuck-convert"
Repository = "https://github.com/Shan-Zhu/shuck-convert"
Issues = "https://github.com/Shan-Zhu/shuck-convert/issues"

[project.scripts]
shuck-convert = "shuck_convert.cli:main"

[project.entry-points."mcp.servers"]
shuck-convert = "shuck_convert.server:mcp"

[tool.hatch.build.targets.wheel]
packages = ["src/shuck_convert"]

[tool.pytest.ini_options]
testpaths = ["tests"]
